Quick work-around for some variant pokemon causing issues (alola, galar etc); fixed PokemonId conversion list.
Passing the pokemon species id to the convertPokemonId() function works in some cases (getting the numeral id for the correct sprite), but doesn't work when you want to get the string pokemon name/identifier that pokeapi uses from the I2P map back. That name also wasn't returned in the proper format for variant pokemon.